Revert About redesign, move contributors to Testing
