+++ client/src/pages/Testing.tsx
@@ -1,7 +1,8 @@
-import {ReactNode, useContext, useState} from 'react';
+import {ReactNode, useContext, useEffect, useState} from 'react';
 
 // Components
 import CenteredMessage from '../components/layout/CenteredMessage';
+import OutlineButton, {DangerOutlineButton, SuccessOutlineButton} from '../components/layout/OutlineButton';
 import Period from '../components/schedule/Period';
 import Loading from '../components/layout/Loading';
 import WIP from '../components/layout/WIP';
@@ -61,15 +62,45 @@ export default function Testing() {
     const userData = useContext(UserDataContext);
     const currTime = useContext(CurrentTimeContext);
 
+    const [contributors, setContributors] = useState<any[] | null>(null);
+
+    useEffect(() => {
+        fetch('https://api.github.com/repos/GunnWATT/watt/contributors')
+            .then(res => res.json())
+            .then(json => setContributors(json))
+    }, [])
+
     return (
         <div className="container py-6">
-            <h1>Super Secret Testing Facility</h1>
-            <p>
+            <h1 className="mb-4">Super Secret Testing Facility</h1>
+
+            <p className="mb-4">
                 Congratulations! You found the super secret testing area for Gunn WATT!
                 Experiments and other potential features will live here until they are accepted or rejected,
                 and components that only trigger conditionally (like loading screens for fetched content or period
                 components in a specific state) will stay here permanently for convenience in testing.
             </p>
+
+            <p className="mb-6">
+                While we're here, I'd like to extend a thank you to all of our GitHub contributors, testers, and advisors,
+                who dedicated their time and expertise to work on this project. Your bug fixes and design advice guides
+                the continued improvement of WATT as it heads into the future. The following people made WATT possible:
+            </p>
+            {contributors && (
+                <div className="flex flex-wrap justify-center gap-3 px-4 mb-4">
+                    {contributors.map((c) => (
+                        <Contributor name={c.login} href={c.html_url} src={c.avatar_url} />
+                    ))}
+                </div>
+            )}
+            <div className="flex flex-wrap justify-center gap-x-4 gap-y-2 secondary font-light text-sm px-4 mb-4">
+                <span>Brandon Chung</span>
+                <span>Sean Yen</span>
+                <span>Timothy Herchen</span>
+                <span>Saumya Singhal</span>
+                <span>Mylie Rodrigo</span>
+                <span>Victor Dang</span>
+            </div>
             <hr/>
 
             <main className="flex flex-col gap-4">
@@ -92,6 +123,12 @@ export default function Testing() {
 
                 <ContentBox>
                     <SgySignInBtn />
+                </ContentBox>
+
+                <ContentBox className="flex gap-2 justify-center">
+                    <OutlineButton>Add to my list</OutlineButton>
+                    <DangerOutlineButton>Disable Schoology</DangerOutlineButton>
+                    <SuccessOutlineButton>Looks good!</SuccessOutlineButton>
                 </ContentBox>
 
                 <ContentBox>
@@ -168,10 +205,19 @@ export default function Testing() {
     )
 }
 
-function ContentBox(props: {children: ReactNode}) {
+function ContentBox(props: {children: ReactNode, className?: string}) {
     return (
-        <section className="p-5 rounded-lg bg-content dark:bg-content-dark shadow-lg">
+        <section className={'p-5 rounded-lg bg-content dark:bg-content-dark shadow-lg' + (props.className ? ` ${props.className}` : '')}>
             {props.children}
         </section>
+    )
+}
+
+type ContributorProps = {name: string, href: string, src: string};
+function Contributor(props: ContributorProps) {
+    return (
+        <a href={props.href} target="_blank" rel="noopener noreferrer" className="rounded-full overflow-clip hover:ring-4 hover:ring-tertiary dark:hover:ring-tertiary-dark transition duration-100">
+            <img src={props.src} alt={props.name} className="h-12 w-12" />
+        </a>
     )
 }
